Improvements after code review
